+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to the "aiXcoder" extension will be documented in this file.
 
+## [0.2.3] - 2019-11-18
+
+### Added
+- Progress display for launching local server.
+
+### Changed
+
+### Fixed
+- Fix a bug that prevents local server launching automatically.
+
 ## [0.2.2] - 2019-11-18
 
 ### Added